+++ src/main/java/com/telepathicgrunt/bumblezone/entities/mobs/BeehemothEntity.java
@@ -126,7 +126,7 @@ public class BeehemothEntity extends TamableAnimal implements FlyingAnimal {
     }
 
     public void addFriendship(Integer deltaFriendship) {
-        this.entityData.set(FRIENDSHIP, getFriendship() + deltaFriendship);
+        this.entityData.set(FRIENDSHIP, Math.min(1000, getFriendship() + deltaFriendship));
     }
 
     @Override