+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Beehemoth speed can not be changed by config in the world's serverconfig folder 
 Holding down Space (jump button) while riding Beehemoth will make the bee fly upward.
 
 Fixed bug where you could feed Beehemoth to surpass the 1000 point limit on friendship.
+ Capped at 1000 not and cannot be lowered below -100 too.
 
 
 ### **(V.3.2.3 Changes) (1.17.1 Minecraft)**